+++ mocking/hardware/vendors.h
@@ -39,6 +39,7 @@
     #include "../stm32/gpio_stm32_fake.h"
     #include "../stm32/i2c_stm32_fake.h"
     #include "../stm32/spi_stm32_fake.h"
+    #include "../stm32/sdio_stm32_fake.h"
     #include "../stm32/exti_stm32_fake.h"
     #include "../stm32/timer_stm32_fake.h"
     #include "../stm32/usart_stm32_fake.h"